dirtyheroku/requirements.txt
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Pillow==8.3.1
psycopg2==2.9.1
PyJWT==2.1.0
cloudinary==1.26.0
dj-database-url==0.5.0
gunicorn==20.1.0
python-decouple==3.4
whitenoise==5.3.0
django-admin-interface